feat: adding the settings to theme settings
