+++ src/components/Tenants/AppSettings/NotificationsSettings/index.tsx
@@ -1,10 +1,10 @@
 import { Divider } from 'antd';
 import { useTranslation } from 'react-i18next';
-import { CardEditable } from '../../../../../components/CardEditable';
-import { FormRadioGroupField } from '../../../../../components/FormRadioGroupField';
-import { FormSwitchField } from '../../../../../components/FormSwitchField';
-import { useSingleTenantData } from '../../../../../hooks/useSingleTenantData';
-import { useTenantAdminDataMutation } from '../../../../../hooks/useTenantAdminDataMutation.hook';
+import { CardEditable } from '../../../CardEditable';
+import { FormRadioGroupField } from '../../../FormRadioGroupField';
+import { FormSwitchField } from '../../../FormSwitchField';
+import { useSingleTenantData } from '../../../../hooks/useSingleTenantData';
+import { useTenantAdminDataMutation } from '../../../../hooks/useTenantAdminDataMutation.hook';
 import styles from './styles.module.scss';
 
 export const NotificationsSettings = ({ tenantId }: { tenantId: string }) => {
