+++ src/components/Tenants/GeneralSettings/components/TypeOfLanguage/styles.module.scss
