+++ src/components/Tenants/LegalSettings/components/LegalText/styles.module.scss
