+++ src/components/Tenants/LegalSettings/styles.module.scss
