+++ src/pages/TenantSettings/index.tsx
@@ -1,10 +1,12 @@
 import { Outlet } from 'react-router';
 import { Page } from '../../components/Page';
+import { useAppConfigContext } from '../../context/useAppConfig';
 import { PermissionAction } from '../../enums/PermissionAction';
 import { Resource } from '../../enums/Resource';
 import { useUserPermissions } from '../../hooks/useUserPermission';
 
 export const TenantSettingsLayout = () => {
+    const { settings } = useAppConfigContext();
     const { can } = useUserPermissions();
 
     return (
@@ -20,6 +22,12 @@ export const TenantSettingsLayout = () => {
                     can(PermissionAction.Update, Resource.LegalText) && {
                         to: '/admin/theme-settings/legal',
                         titleKey: 'settings.subhead.legal',
+                    },
+                    can(PermissionAction.Update, Resource.Tenant) && {
+                        to: '/admin/theme-settings/app-settings',
+                        titleKey: `tenants.edit.tabs.${
+                            settings.multitenancyWithSingleDomainEnabled ? 'globalSettings' : 'appSettings'
+                        }`,
                     },
                 ]}
             />
