+++ src/pages/Tenants/Edit/LegalSettings/index.tsx
@@ -1,5 +1,5 @@
 import { useParams } from 'react-router';
-import { LegalSettings } from '../../../TenantSettings/LegalSettings';
+import { LegalSettings } from '../../../../components/Tenants/LegalSettings';
 
 export const SingleLegalSettings = () => {
     const { id } = useParams();
