+++ src/pages/Tenants/Edit/ThemeSettings/index.tsx
@@ -1,5 +1,5 @@
 import { useParams } from 'react-router';
-import { GeneralSettings } from '../../../TenantSettings/GeneralSettings';
+import { GeneralSettings } from '../../../../components/Tenants/GeneralSettings';
 
 export const TenantThemeSettings = () => {
     const { id } = useParams();
